+++ modules/braille/liblouis-utils/src/test/java/LiblouisCoreTest.java
@@ -125,7 +125,7 @@ public class LiblouisCoreTest extends AbstractTest {
 	
 	@Test
 	public void testGetTranslatorFromQuery3() {
-		provider.withContext(messageBus).get(query("(locale:foo_BA)")).iterator().next();
+		provider.withContext(messageBus).get(query("(document-locale:foo_BA)")).iterator().next();
 	}
 	
 	@Test(expected=NoSuchElementException.class)