+++ modules/braille/texhyph-utils/src/main/java/org/daisy/pipeline/braille/tex/impl/TexHyphenatorDotifyImpl.java
@@ -93,8 +93,8 @@ public class TexHyphenatorDotifyImpl extends AbstractTransformProvider<TexHyphen
 			return fromNullable(get(URLs.asURI(v))); }
 		Locale locale; {
 			String loc;
-			if (q.containsKey("locale"))
-				loc = q.removeOnly("locale").getValue().get();
+			if (q.containsKey("document-locale"))
+				loc = q.removeOnly("document-locale").getValue().get();
 			else
 				loc = "und";
 			try {